made a new component react-native-vector-icons and styles
